refactor: extract shared higress helpers into helpers.ts
- Move getSessionCookie, isRecord, unwrapData, maskProvider to helpers.ts
- Update all 4 higress route files to import from helpers
- Eliminates 4x duplication of identical helper functions
- Reduces total route code by ~40 lines

Closes #35 TASK-002
